added the stock management and the user empty cart while creation, some other bugs solved

+++ models/poducts.go
@@ -18,7 +18,7 @@ type Product struct {
 	Image         string     `gorm:"not null"`
 	Weight        float64    `gorm:"not null"` // Required
 	Categories    []Category `gorm:"many2many:product_categories;"`
-	Stock         float64
+	Stock         int
 	CreatedAt     time.Time
 	UpdatedAt     time.Time
 	DeletedAt     gorm.DeletedAt `gorm:"index"`
